refactor(celery): Eliminate ambiguity in import scripts
I was getting an error about local imports and partial imports and
circular imports, and it was all because we had a file named
celery.py. I'm not sure if this is a py3 thing or a PyCharm thing (b/c
it doesn't happen in other contexts, I think), but anyway, the fix was
simple: Stop having a file called celery.py.
