+++ cl/__init__.py
@@ -1,3 +1,3 @@
 # This will make sure the app is always imported when
 # Django starts so that shared_task will use this app.
-from .celery import app as celery_app
+from .celery_init import app as celery_app
