+++ cl/stats/views.py
@@ -1,6 +1,6 @@
 from django.http import HttpResponse
 
-from cl.celery import fail_task
+from cl.celery_init import fail_task
 from cl.lib.redis_utils import make_redis_interface
